requirements.txt
altair==5.5.0
apturl==0.5.2
attrs==24.3.0
awscli==1.22.34
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.4
botocore==1.23.34
Brlapi==0.8.3
cachetools==5.5.0
certifi==2024.12.14
chardet==4.0.0
charset-normalizer==3.4.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
duplicity==0.8.21
exceptiongroup==1.2.2
faiss-cpu==1.9.0.post1
fasteners==0.14.1
filelock==3.16.1
fsspec==2024.12.0
future==0.18.2
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httplib2==0.20.2
huggingface-hub==0.27.0
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.5
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
narwhals==1.19.1
netifaces==0.11.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==5.29.2
psutil==5.9.1
ptyprocess==0.7.0
pyarrow==18.1.0
pyasn1==0.4.8
pycairo==1.20.1
pycups==2.0.1
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
referencing==0.35.1
regex==2024.11.6
reportlab==3.6.8
requests==2.32.3
rich==13.9.4
roman==3.3
rpds-py==0.22.3
rsa==4.8
s3transfer==0.5.0
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
SecretStorage==3.3.1
selenium==4.27.1
sentence-transformers==3.3.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
ssh-import-id==5.11
streamlit==1.41.1
sympy==1.13.1
systemd-python==234
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.1
trio==0.27.0
trio-websocket==0.11.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
